+++ app/src/main/java/com/example/buyinglistocr/model/ItemManager.java
@@ -8,6 +8,7 @@ import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
+import com.example.buyinglistocr.util.RequestHandler;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -31,7 +32,7 @@ public class ItemManager {
      * @param item - The item
      * @return - The id of the item
      */
-    public long add(final Item item) {
+    public int add(final Item item) {
 
         StringRequest stringRequest = new StringRequest(Request.Method.POST, "http://51.83.70.93/android/BuyingListOCR/ListIndex.php",
